requirements.txt
Django==4.2.*
djangorestframework==3.15.*
PyMySQL==1.1.*
django-cors-headers==4.4.*
django-environ==0.11.*
python-dotenv==1.0.*
drf-spectacular==0.27.*
whitenoise==6.7.*
markdown==3.7
pygments==2.18.*
bleach==6.1.*
cryptography==44.*
Pillow==11.*
black==24.*
ruff==0.3.*
mypy==1.9.*
django-stubs==4.2.*
